villain-bestiary/requirements.txt
requests
beautifulsoup4
pyyaml
mistralai
